Refactor shared methods
